requirements.txt
requests==2.28.1
requests_toolbelt==0.9.1
Pillow==9.2.0
